requirements.txt
adium-theme-ubuntu==0.3.4
apt-xapian-index==0.45
cffi==1.1.2
chardet==2.0.1
colorama==0.2.5
command-not-found==0.3
cryptography==0.9.1
debtagshw==0.1
decorator==3.4.0
defer==1.0.6
dirspec==13.10
Django==1.8.2
django-filter==0.10.0
django-guardian==1.3
djangorestframework==3.1.3
docopt==0.6.2
duplicity==0.6.23
enum34==1.0.4
html5lib==0.999
httplib2==0.8
idna==2.0
ipaddress==1.0.7
joblib==0.7.1
lockfile==0.8
lxml==3.3.3
Markdown==2.6.2
matplotlib==1.3.1
mongoengine==0.9.0
ngxtop==0.0.2
nose==1.3.1
numpy==1.8.2
oauthlib==0.6.1
oneconf==0.3.7
PAM==0.4.2
pexpect==3.1
Pillow==2.3.0
piston-mini-client==0.7.5
psycopg2==2.4.5
pyasn1==0.1.7
pycparser==2.14
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pygobject==3.12.0
pymongo==3.0.2
pyOpenSSL==0.13
pyparsing==2.0.3
pyserial==2.6
pysmbc==1.0.14.1
python-apt==0.9.3.5
python-dateutil==1.5
python-debian===0.1.21-nmu2ubuntu2
pytz==2012rc0
pyxdg==0.25
reportlab==3.0
requests==2.2.1
scikit-learn==0.14.1
scipy==0.13.3
sessioninstaller==0.0.0
simplejson==3.3.1
six==1.5.2
software-center-aptd-plugins==0.0.0
system-service==0.1.6
tabulate==0.7.5
Twisted-Core==13.2.0
Twisted-Web==13.2.0
unity-lens-photos==1.0
urllib3==1.7.1
uWSGI==2.0.10
virtualenv==13.0.3
wheel==0.24.0
xdiagnose===3.6.3build2
zope.interface==4.0.5
